VGG-16/requirements.txt
torch
torchvision
numpy
Pillow
scikit-image